My Latest Update

+++ AMS.Model/Models/AmsNeo4JNodeConstraint.cs
@@ -5,7 +5,7 @@ namespace AMS.Model.Models;
 
 public partial class AmsNeo4JNodeConstraint
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
 
     public string? Type { get; set; }
 
@@ -13,7 +13,7 @@ public partial class AmsNeo4JNodeConstraint
 
     public string? Command { get; set; }
 
-    public long? LabelFk { get; set; }
+    public int? LabelFk { get; set; }
 
-    public long? RelTypeFk { get; set; }
+    public int? RelTypeFk { get; set; }
 }
