+++ AMS.Model/Models/AmsNeo4JNodeDepartment.cs
@@ -5,11 +5,11 @@ namespace AMS.Model.Models;
 
 public partial class AmsNeo4JNodeDepartment
 {
-    public long NodeDepartmentId { get; set; }
+    public int NodeDepartmentId { get; set; }
 
-    public long? DepartmentId { get; set; }
+    public int? DepartmentId { get; set; }
 
-    public long? CmsClassId { get; set; }
+    public int? CmsClassId { get; set; }
 
-    public long? NodeId { get; set; }
+    public int? NodeId { get; set; }
 }
