+++ AMS.Model/Models/AmsNeo4JNodeLabelQuery.cs
@@ -5,7 +5,7 @@ namespace AMS.Model.Models;
 
 public partial class AmsNeo4JNodeLabelQuery
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
 
     public string? Name { get; set; }
 
@@ -13,7 +13,7 @@ public partial class AmsNeo4JNodeLabelQuery
 
     public string? Description { get; set; }
 
-    public long? LabelId { get; set; }
+    public int? LabelId { get; set; }
 
     public string? RelatedLabels { get; set; }
