+++ AMS.Model/Models/AmsNeo4Jdepartment.cs
@@ -5,9 +5,9 @@ namespace AMS.Model.Models;
 
 public partial class AmsNeo4JDepartment
 {
-    public long DepartmentId { get; set; }
+    public int DepartmentId { get; set; }
 
     public string? Title { get; set; }
 
-    public long? ProjectFk { get; set; }
+    public int? ProjectFk { get; set; }
 }
